requirements.txt
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
bcrypt==3.1.4
boto3==1.9.33
botocore==1.12.33
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
daphne==2.2.2
dataclasses==0.4
Django==2.0.7
django-bootstrap3==9.1.0
django-cors-headers==2.4.0
django-extensions==1.9.9
django-filter==2.0.0
django-storages==1.7.1
django-test-addons==1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djongo==1.2.30
dnspython==1.15.0
docutils==0.14
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.3
mongoengine==0.15.0
msgpack==0.5.6
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.5.3
pymodm==0.4.0
pymongo==3.6.0
PyNaCl==1.3.0
pypiwin32==223
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.5
pywin32==224
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
sshtunnel==0.1.4
twilio==6.19.2
Twisted==18.7.0
txaio==18.8.1
typing==3.6.4
urllib3==1.24
zope.interface==4.5.0
